fix: try/catch safety for parsing failures
When failing to parse git's output multiple exceptions are thrown
because `l:matched` may not have 5 entries.

This occurs during a rebase for instance where the regexp is unmatched.

I don't think this plugin should scan for in-progress rebases but if you
wanted to do it, parsing `git status`'s output would do the trick (at
the cost of another call to git).

Fixes https://github.com/macthecadillac/lightline-gitdiff/issues/1